finish of simple merge

+++ src/util/imglib.py
@@ -290,7 +290,7 @@ def RGBtoYCbCr(rgbArray):
 
     r,g,b = rgbArray[:,:,0], rgbArray[:,:,1], rgbArray[:,:,2]
     # transMatrix = np.array()
-    outArray = np.empty_like(rgbArray)
+    outArray = np.empty_like(rgbArray).astype(float)
 
     outArray[:,:,0] = 0.299*r + 0.587*g + 0.114*b
     outArray[:,:,1] = -0.169*r + -0.331*g + 0.5*b
